UC/requirements.txt
gunicorn
pandas
openai
django
mysql-connector-python
mysqlclient
python-dotenv
django-cors-headers
